chore: move application settings to header
